Former-commit-id: 95304ccffdda41df35619b6d16847a55279843bd

+++ src/localization/position_solver.h
@@ -8,7 +8,7 @@
 namespace localization {
 
 constexpr double ktag_size = 0.1651;  // meters
-const std::vector<cv::Point3f> kapriltag_corners = {
+const std::vector<cv::Point3d> kapriltag_corners = {
     {-ktag_size / 2, ktag_size / 2, 0},
     {ktag_size / 2, ktag_size / 2, 0},
     {ktag_size / 2, -ktag_size / 2, 0},
